meh
cleanup response

cleanup request and context

yay

make request and response easier

remove the merging of descriptors
